# Speedup polymod in ~15x times