--- a/.gitignore
+++ b/.gitignore
@@ -1,3 +1,5 @@
+node_modules
+
 # Logs
 logs
 *.log