--- a/benchmarks/package.json
+++ b/benchmarks/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "cashaddrjs-benchmark",
+  "version": "0.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "MIT",
+  "dependencies": {
+    "xorshift.js": "^1.0.0"
+  }
+}