Benchmark result:
```
$ SEED=64403a4d05e393ac93683b1ff825f342 node index.js 
Seed: 64403a4d05e393ac93683b1ff825f342
Spent 5.000023416s for 256735 encode/decode: 51346.760 op/s

$ SEED=64403a4d05e393ac93683b1ff825f342 node index.js 
Seed: 64403a4d05e393ac93683b1ff825f342
Spent 5.000157577s for 15747 encode/decode: 3149.301 op/s
```