@ealmansi 2018-12-03
Compare the function [polymod](https://github.com/bitcoincashjs/cashaddrjs/blob/master/src/cashaddr.js#L284) against the official spec: https://github.com/bitcoincashorg/bitcoincash.org/blob/master/spec/cashaddr.md#checksum.

The code in this library is essentially a straightforward translation of the code snippet provided in the spec into JavaScript, using a well-tested library for dealing with big integers.

With this PR's version of the [polymod](https://github.com/fanatid/cashaddrjs/blob/2d4385309c4f2ac0ab2bd9e49b72a3bf6425373d/src/cashaddr.js#L285) function, the process of verifying that the library matches the spec becomes less trivial.

For speed, you can use something like [this](https://github.com/deadalnix/cashaddressed), or something like [this](https://github.com/schancel/cashaddr-converter).