@fanatid 2019-04-28
For notes, with node v12:
```
// bitcoincashjs/cashaddrjs
$ SEED=64403a4d05e393ac93683b1ff825f342 node index.js 
Seed: 64403a4d05e393ac93683b1ff825f342
Spent 5.000286579s for 17785 encode/decode: 3556.796 op/s
// fanatid/cashaddrjs#performance-bigint
$ SEED=64403a4d05e393ac93683b1ff825f342 node index.js 
Seed: 64403a4d05e393ac93683b1ff825f342
Spent 5.000098226s for 34732 encode/decode: 6946.264 op/s
// fanatid/cashaddrjs#performance
$ SEED=64403a4d05e393ac93683b1ff825f342 node index.js 
Seed: 64403a4d05e393ac93683b1ff825f342
Spent 5.000026432s for 373435 encode/decode: 74686.605 op/s
```